requirements.txt
flask
flask-session
pandas
python-dotenv
openai
langchain
langchain-community
